requirements.txt
configobj==5.0.6
pre-commit==2.14.1
pyotp==2.3.0
python-dateutil==2.8.1
python-dotenv==0.11.0
selenium==3.141.0
